Project.py/requirements.txt
matplotlib 
pandas
numpy
streamlit
plotly.express
folium